Added sequence numbers to UDP packets
Appended seqNum and a delimiter to data sent to client over UDP; client side parsed the data for the delimiter and added the rest of the string to be stored.  It's not pretty, but it seems to work.  So far I haven't ran into the address issue of running another client while the server is still active.